feature: create useFormatDate hook
