+++ src/pages/admin/ProjectsPage/EditProjectPage.tsx
@@ -1,59 +1,45 @@
 import Typography from "../../../components/Typography/Typography.tsx";
+import { NavLink, useParams } from "react-router-dom";
 import IconClose from "../../../components/AdminPanel/IconButtons/IconClose.tsx";
 import { Form, Formik } from "formik";
+import { createCard, getCardById } from "../../../api/CardsApi.ts";
 import AdminInput from "../../../components/AdminPanel/Input/AdminInput.tsx";
 import CitiesDropDown from "../../../components/AdminPanel/DropwDown/CitiesDropDown.tsx";
 import CategoryDropDown from "../../../components/AdminPanel/DropwDown/CategoryDropDown.tsx";
 import ImageInput from "../../../components/ImageCroper/ImageInput.tsx";
 import Switch from "../../../components/Switch/Switch.tsx";
-import * as Yup from "yup";
 import Button from "../../../components/Button/Button.tsx";
 import CustomCalendar from "../../../components/AdminPanel/Calendar/CustomCalendar.tsx";
-import { useState } from "react";
-import { NavLink } from "react-router-dom";
-import { createCard } from "../../../api/CardsApi.ts";
+import { useEffect, useState } from "react";
+import { validationSchema } from "./validationShema.ts";
+import { useFormatDate } from "../../../hooks/useFormatDate.tsx";
 
-const AddProjectPage = () => {
+const EditProjectPage = () => {
+	const [isOpen, setIsOpen] = useState(false);
+	const [cardInfo, setCardInfo] = useState();
+	const { id } = useParams<string>();
+	const formatDate = useFormatDate()
+	console.log(cardInfo);
 
-	const [isOpen, setIsOpen] = useState(false)
+	useEffect(() => {
+		if (id) {
+			getCardById(id)
+				.then(data => setCardInfo(data));
+		}
+	}, [id]);
 
-	const validationSchema = Yup.object({
-		title: Yup.string()
-			.required("Поле обов'язкове до заповнення"),
-		url: Yup.string()
-			.required("Поле обов'язкове до заповнення"),
-		description: Yup.string()
-			.required("Поле обов'язкове до заповнення"),
-		city: Yup.string()
-			.required("Місто та країна обовязкове"),
-		country: Yup.string()
-			.required("Місто та країна обовязкове"),
-		image: Yup.mixed()
-			.required("Поле обов'язкове до заповнення"),
-		publication: Yup.string()
-			.required("Поле обов'язкове до заповнення"),
-		category: Yup.string()
-			.required("Поле обов'язкове до заповнення"),
-	});
-
-	const currentDate = new Date();
-	const formattedCurrentDate = currentDate.toLocaleDateString('uk-UA', {
-		day: '2-digit',
-		month: '2-digit',
-		year: 'numeric'
-	});
 
 	return (
 		<>
 			<div className={"bg-grey100 "}>
 				<div className={"px-[36px] pt-[38px] pb-[38px] pr-[80px] h-[118px] flex justify-between"}>
 					<Typography variant={"h3"} component={"h3"} className={"text-white"}>Проєкти</Typography>
-					<NavLink to={'/admin/projects/'}><IconClose /></NavLink>
+					<NavLink to={"/admin/projects/"}><IconClose /></NavLink>
 
 				</div>
 			</div>
 			<div className={"pt-[48px] pl-[34px] pr-[80px] pb-[128px] bg-grey30 h-[100vh]"}>
-				<Typography variant={"h4"} component={"h4"} className={"text-black mb-5"}>Додати проєкт</Typography>
+				<Typography variant={"h4"} component={"h4"} className={"text-black mb-5"}>Редагувати проєкт</Typography>
 				<Formik
 					initialValues={{
 						title: "",
@@ -63,7 +49,7 @@ const AddProjectPage = () => {
 						country: "",
 						image: "",
 						isEnabled: true,
-						publication: formattedCurrentDate,
+						publication: formatDate,
 						category: "",
 					}}
 					validationSchema={validationSchema}
@@ -157,7 +143,7 @@ const AddProjectPage = () => {
 
 													<div className={"flex items-center text-[14px] relative"}>
 														Дата публікації:
-														<button onClick={() => setIsOpen(!isOpen)} className={'underline cursor-pointer p-2'}>
+														<button onClick={() => setIsOpen(!isOpen)} className={"underline cursor-pointer p-2"}>
 															{values.publication}
 														</button>
 													</div>
@@ -168,10 +154,10 @@ const AddProjectPage = () => {
 												<CustomCalendar
 													setIsOpen={setIsOpen}
 													onValueSelected={(date) => {
-														if(date){
-															setFieldValue("publication", date)
+														if (date) {
+															setFieldValue("publication", date);
 														} else {
-															setFieldValue("publication", formattedCurrentDate)
+															setFieldValue("publication", formattedCurrentDate);
 														}
 													}}
 												/>}
@@ -188,4 +174,4 @@ const AddProjectPage = () => {
 	);
 };
 
-export default AddProjectPage;
+export default EditProjectPage;
